requirements_frozen.txt
appdirs==1.4.3

atomicwrites==1.3.0

attrs==19.1.0

black==19.3b0

certifi==2019.6.16

chardet==3.0.4

Click==7.0

codecov==2.0.15

coverage==4.5.4

docutils==0.15.2

effect==0.12.0

entrypoints==0.3

fancycompleter==0.8

flake8==3.7.8

flake8-debugger==3.1.0

flake8-isort==2.7.0

flake8-quotes==2.1.0

flake8-unused-arguments==0.0.3

flit==1.3

idna==2.8

importlib-metadata==0.21

intreehooks==1.0

isort==4.3.21

Jinja2==2.10.1

MarkupSafe==1.1.1

mccabe==0.6.1

mkdir==2019.4.13

more-itertools==7.2.0

mypy==0.720

mypy-extensions==0.4.1

nix-prefetch-github==2.3.1

orderdict==2019.4.13

packaging==19.1

Parsley==1.3

pdbpp==0.10.0

pluggy==0.13.0

public==2019.4.13

py==1.8.0

pycodestyle==2.5.0

pyflakes==2.1.1

Pygments==2.4.2

pyparsing==2.4.2

pytest==5.1.2

pytest-cov==2.7.1

pytest-runner==5.1

pytoml==0.1.21

requests==2.22.0

setupcfg==2019.4.13

setuptools-scm==3.3.3

six==1.12.0

testfixtures==6.10.0

toml==0.10.0

typed-ast==1.4.0

typing-extensions==3.7.4

urllib3==1.25.3

values==2019.4.13

wcwidth==0.1.7

wmctrl==0.3

write==2019.4.13

zipp==0.6.0